upsetPeaks

upsetPeaks

upsetPeaks(
  peaks,
  conds = names(peaks),
  conds_order = conds,
  order.by = "freq",
  mainbar.y.label = "Intersect size",
  sets.x.label = "Set size",
  ...
)

Arguments

peaks
List of dataframes with the genomic coordinates of the regions to overlap. Dataframes must contain the columns seqnames, start, end.

conds
Character with the same length as the 'peaks' list. The names that are given to the diferente objects in the 'peaks' list. Default: 'names(peaks)'

conds_order
Character of the same length as 'conds'. Same values as in 'conds' but with the desired order of priority. Default: 'conds'.

order.by
Same as in UpSetR::upset(). One of "freq", "degree" or both.

mainbar.y.label
Same as in UpSetR::upset().

sets.x.label
Same as in UpSetR::upset().

...
Further arguments to be passed through 'UpSetR::upset()'.

title
(Not available yet) Character of length 1 or NULL. Title of the plot shown at the top. Passed through grid.arrange(). If NULL, title is not shown. Default: NULL

See also

Author

amitjavilaventura

This function calls 'getVennCounts()' and draws and UpSet plot using the UpSetR package.